Add user login using Google Sign In and navigate to UserPreferencesScreen (using temporary Screen)
