--- app/src/main/java/com/ch2ps38/nutrimate/ui/navigation/NavigationItem.kt
@@ -1,9 +1,0 @@
-package com.ch2ps38.nutrimate.ui.navigation
-
-import androidx.compose.ui.graphics.vector.ImageVector
-
-data class NavigationItem(
-    val title: String,
-    val icon: ImageVector,
-    val screen: Screen,
-)
