--- app/src/main/java/com/ch2ps38/nutrimate/ui/navigation/Screen.kt
@@ -1,9 +1,0 @@
-package com.ch2ps38.nutrimate.ui.navigation
-
-sealed class Screen(val route: String){
-    object  Home : Screen ("home")
-    object Menu : Screen ("menu")
-    object Planner : Screen ("planner")
-    object Reminder : Screen ("reminder")
-    object Profile : Screen ("profile")
-}
