+++ app/src/main/java/com/ch2ps385/nutrimate/MainActivity.kt
@@ -1,0 +1,34 @@
+package com.ch2ps385.nutrimate
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.ch2ps385.nutrimate.ui.theme.NutriMateTheme
+import androidx.navigation.compose.navigation
+import androidx.navigation.compose.rememberNavController
+import com.ch2ps385.nutrimate.NutriMateApp
+
+class MainActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContent {
+            NutriMateTheme {
+                // A surface container using the 'background' color from the theme
+                Surface(
+                    modifier = Modifier.fillMaxSize(),
+                    color = MaterialTheme.colorScheme.background
+                ) {
+                    NutriMateApp()
+                }
+            }
+        }
+    }
+}
+
