+++ app/src/main/java/com/ch2ps385/nutrimate/NutriMateApp.kt
@@ -1,0 +1,144 @@
+package com.ch2ps385.nutrimate
+
+import android.app.Activity.RESULT_OK
+import android.content.ContentValues.TAG
+import android.util.Log
+import android.widget.Toast
+import androidx.activity.compose.rememberLauncherForActivityResult
+import androidx.activity.result.IntentSenderRequest
+import androidx.activity.result.contract.ActivityResultContracts
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Scaffold
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.ch2ps385.nutrimate.presentation.screen.auth.signin.GoogleAuthUiClient
+import com.ch2ps385.nutrimate.presentation.screen.auth.signin.SignInScreen
+import com.ch2ps385.nutrimate.presentation.screen.auth.signin.SignInViewModel
+import com.ch2ps385.nutrimate.presentation.screen.user.preferences.UserPreferenceScreen
+import com.ch2ps385.nutrimate.ui.component.other.BottomBar
+import com.ch2ps385.nutrimate.ui.component.other.Toolbar
+import com.ch2ps385.nutrimate.ui.navigation.Screen
+import com.google.android.gms.auth.api.identity.Identity
+import kotlinx.coroutines.launch
+
+@Composable
+fun NutriMateApp(
+    modifier : Modifier = Modifier,
+    navController : NavHostController = rememberNavController(),
+){
+    val navBackStackEntry by navController.currentBackStackEntryAsState()
+    val currentRoute = navBackStackEntry?.destination?.route
+    val applicationContext = LocalContext.current
+    val googleAuthUiClient by lazy {
+        GoogleAuthUiClient(
+            context = applicationContext,
+            oneTapClient = Identity.getSignInClient(applicationContext)
+        )
+    }
+    val coroutineScope = rememberCoroutineScope()
+
+    Scaffold(
+        topBar = {
+                 if(currentRoute != Screen.SignIn.route){
+                     Toolbar()
+                 }
+                 },
+        bottomBar = {
+            if(currentRoute != Screen.UserPreferences.route && currentRoute != Screen.SignIn.route){
+                BottomBar(navController)
+            }
+        }
+    ) {innerPadding ->
+        NavHost(
+            navController = navController,
+            startDestination = Screen.SignIn.route,
+            modifier = Modifier.padding(innerPadding)
+        ){
+            composable(Screen.SignIn.route){
+                val viewModel = viewModel<SignInViewModel>()
+                val state by viewModel.state.collectAsStateWithLifecycle()
+                LaunchedEffect(key1 = Unit) {
+                    if(googleAuthUiClient.getSignedInUser() != null) {
+                        navController.navigate("userpreferences")
+                    }
+                }
+
+                // activity result after user selecting the gmail user profile
+                val launcher = rememberLauncherForActivityResult(
+                    contract = ActivityResultContracts.StartIntentSenderForResult(),
+                    onResult = { result ->
+                        Log.d(TAG, "onCreate: Result ${result.resultCode}")
+                        if(result.resultCode == RESULT_OK) {
+                            coroutineScope.launch {
+                                val signInResult = googleAuthUiClient.signInWithIntent(
+                                    intent = result.data ?: return@launch
+                                )
+                                viewModel.onSignInResult(signInResult)
+                            }
+                        }
+                    }
+                )
+
+
+                // Updating user/UI after sign in successful
+                LaunchedEffect(key1 = state.isSignInSuccessful) {
+                    if(state.isSignInSuccessful) {
+                        Toast.makeText(
+                            applicationContext,
+                            "Sign in successful",
+                            Toast.LENGTH_LONG
+                        ).show()
+                        navController.navigate("userpreferences")
+                        viewModel.resetState()
+                    }
+                }
+
+                //Initial screen with Sign in button
+                SignInScreen(
+                    state = state,
+                    onSignInClick = {
+                        coroutineScope.launch {
+                            val signInIntentSender = googleAuthUiClient.signIn()
+                            launcher.launch(
+                                IntentSenderRequest.Builder(
+                                    signInIntentSender ?: return@launch
+                                ).build()
+                            )
+                        }
+                    }
+                )
+            }
+            composable(Screen.UserPreferences.route){
+                UserPreferenceScreen(
+                    userData = googleAuthUiClient.getSignedInUser(),
+                    onSignOut = {
+                        coroutineScope.launch {
+                            googleAuthUiClient.signOut()
+                            Toast.makeText(
+                                applicationContext,
+                                "Signed out",
+                                Toast.LENGTH_LONG
+                            ).show()
+
+                            navController.popBackStack()
+                        }
+                    }
+                )
+            }
+
+        }
+
+    }
+
+}
