+++ app/src/main/java/com/ch2ps385/nutrimate/presentation/screen/auth/signin/GoogleAuthUiClient.kt
@@ -1,0 +1,93 @@
+package com.ch2ps385.nutrimate.presentation.screen.auth.signin
+
+import android.content.Context
+import android.content.Intent
+import android.content.IntentSender
+import com.ch2ps385.nutrimate.data.SignInResult
+import com.ch2ps385.nutrimate.data.UserData
+import com.google.android.gms.auth.api.identity.BeginSignInRequest
+import com.google.android.gms.auth.api.identity.BeginSignInRequest.GoogleIdTokenRequestOptions
+import com.google.android.gms.auth.api.identity.SignInClient
+import com.google.firebase.Firebase
+import com.google.firebase.auth.GoogleAuthProvider
+import com.google.firebase.auth.auth
+import kotlinx.coroutines.tasks.await
+import java.util.concurrent.CancellationException
+
+class GoogleAuthUiClient(
+    private val context : Context,
+    private val oneTapClient : SignInClient
+    ) {
+    private val auth = Firebase.auth
+
+    suspend fun signIn():IntentSender?{
+        val result = try{
+            oneTapClient.beginSignIn(
+                buildSignInRequest()
+            ).await()
+        }catch (e : Exception){
+            e.printStackTrace()
+            if(e is CancellationException) throw e
+            null
+        }
+        return result?.pendingIntent?.intentSender
+    }
+
+    suspend fun signInWithIntent(intent: Intent) : SignInResult {
+        val credential = oneTapClient.getSignInCredentialFromIntent(intent)
+        val googleIdToken = credential.googleIdToken
+        val googleCredentials = GoogleAuthProvider.getCredential(googleIdToken, null)
+        return try{
+            val user = auth.signInWithCredential(googleCredentials).await().user
+            SignInResult(
+                data = user?.run {
+                    UserData(
+                        userId = uid,
+                        username = displayName,
+                        profilePictureUrl = photoUrl?.toString()
+                    )
+                },
+                errorMessage = null
+            )
+        } catch (e: Exception){
+            e.printStackTrace()
+            if (e is CancellationException) throw e
+            SignInResult(
+                data = null,
+                errorMessage = e.message
+            )
+        }
+    }
+
+    suspend fun signOut(){
+        try {
+            oneTapClient.signOut().await()
+            auth.signOut()
+        } catch (e: Exception){
+            e.printStackTrace()
+            if (e is CancellationException) throw e
+        }
+    }
+
+    fun getSignedInUser(): UserData? = auth.currentUser?.run {
+        UserData(
+            userId = uid,
+            username = displayName,
+            profilePictureUrl = photoUrl?.toString()
+        )
+    }
+
+
+    private fun buildSignInRequest(): BeginSignInRequest {
+        return BeginSignInRequest.Builder()
+            .setGoogleIdTokenRequestOptions(
+                GoogleIdTokenRequestOptions.builder()
+                    .setSupported(true)
+                    .setFilterByAuthorizedAccounts(false)
+                    .setServerClientId("23383092402-dm8fmcqni6bpcp8g7n9r57tne33h90mo.apps.googleusercontent.com")
+                    .build()
+            )
+            .setAutoSelectEnabled(true)
+            .build()
+    }
+}
