+++ app/src/main/java/com/ch2ps385/nutrimate/presentation/screen/auth/signin/SignInState.kt
@@ -1,0 +1,8 @@
+package com.ch2ps385.nutrimate.presentation.screen.auth.signin
+
+import java.lang.Error
+
+data class SignInState(
+    val isSignInSuccessful : Boolean = false,
+    val signInError: String? = null
+)
