+++ app/src/main/java/com/ch2ps385/nutrimate/presentation/screen/auth/signin/SignInViewModel.kt
@@ -1,0 +1,23 @@
+package com.ch2ps385.nutrimate.presentation.screen.auth.signin
+
+import androidx.lifecycle.ViewModel
+import com.ch2ps385.nutrimate.data.SignInResult
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.flow.update
+
+class SignInViewModel: ViewModel() {
+    private val _state = MutableStateFlow(SignInState())
+    val state = _state.asStateFlow()
+
+    fun onSignInResult(result: SignInResult){
+        _state.update { it.copy(
+            isSignInSuccessful = result.data != null,
+            signInError = result.errorMessage
+        ) }
+    }
+
+    fun resetState(){
+        _state.update { SignInState() }
+    }
+}
