+++ app/src/main/java/com/ch2ps385/nutrimate/ui/component/other/ToolBar.kt
@@ -1,0 +1,58 @@
+package com.ch2ps385.nutrimate.ui.component.other
+
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.*
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import com.ch2ps385.nutrimate.R
+import com.ch2ps385.nutrimate.ui.theme.NutriMateTheme
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun Toolbar(
+    title: String = stringResource(id = R.string.app_name),
+    modifier: Modifier = Modifier,
+    onBackClick: () -> Unit = {}
+) {
+    TopAppBar(
+        modifier = modifier,
+        navigationIcon = {
+            IconButton(
+                onClick = { onBackClick() }
+            ) {
+                Icon(
+                    imageVector = Icons.Default.ArrowBack,
+                    contentDescription = "Back"
+                )
+            }
+        },
+        title = {
+            Row(
+                verticalAlignment = Alignment.CenterVertically,
+                horizontalArrangement = Arrangement.Center
+            ) {
+                Spacer(modifier = Modifier.width(8.dp))
+                Text(
+                    text = title,
+                    fontWeight = FontWeight.Bold,
+                    style = MaterialTheme.typography.titleMedium
+                )
+            }
+        },
+    )
+}
+
+@Preview(showBackground = true)
+@Composable
+fun ToolbarPreview() {
+    NutriMateTheme {
+        Toolbar()
+    }
+}
