+++ app/src/main/java/com/ch2ps385/nutrimate/ui/navigation/NavigationItem.kt
@@ -1,0 +1,9 @@
+package com.ch2ps385.nutrimate.ui.navigation
+
+import androidx.compose.ui.graphics.vector.ImageVector
+
+data class NavigationItem(
+    val title: String,
+    val icon: ImageVector,
+    val screen: Screen,
+)
