+++ app/src/main/java/com/ch2ps385/nutrimate/ui/theme/Color.kt
@@ -1,0 +1,31 @@
+package com.ch2ps385.nutrimate.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
+
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
+
+// Primary Color
+val pSmashedPumpkin = Color(0xFFFE6833)
+val pNaplesYellow = Color(0xFFFBCB5C)
+val pSinopia = Color(0xFFC73A07)
+
+// Secondary Color
+val sUnbleachedSilk = Color(0xFFFFD6C9)
+val sNyanza = Color(0xFFEAFFE0)
+val sBabyPink = Color(0xFFF1CDC1)
+
+// Neutral Color
+// 1 - 6 ( From the Dark to Light )
+val neutralColor1 = Color(0xFF504840)
+val neutralColor2 = Color(0xFF72695E)
+val neutralColor3 = Color(0xFF948A7E)
+val neutralColor4 = Color(0xFFB6ACA0)
+val neutralColor5 = Color(0xFFD8CEC4)
+val neutralColor6 = Color(0xFFFAF2E9)
+val solidWhite = Color(0xFFF8F5F5)
