+++ app/src/main/java/com/ch2ps385/nutrimate/ui/theme/Shape.kt
@@ -1,0 +1,14 @@
+package com.ch2ps385.nutrimate.ui.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Shapes
+import androidx.compose.ui.unit.dp
+
+val Shapes = Shapes(
+    extraSmall = RoundedCornerShape(8.dp),
+    small = RoundedCornerShape(10.dp),
+    medium = RoundedCornerShape(16.dp),
+    large = RoundedCornerShape(20.dp),
+    extraLarge = RoundedCornerShape(80.dp)
+
+)
