+++ app/src/main/res/values/strings.xml
@@ -1,4 +1,6 @@
 <resources>
     <string name="app_name">NutriMate</string>
     <string name="search_hint">Cari apa?</string>
+    <string name="email">Email</string>
+    <string name="password">Password</string>
 </resources>
