+++ app/src/test/java/com/ch2ps385/nutrimate/ExampleUnitTest.kt
@@ -1,0 +1,17 @@
+package com.ch2ps385.nutrimate
+
+import org.junit.Test
+
+import org.junit.Assert.*
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * See [testing documentation](http://d.android.com/tools/testing).
+ */
+class ExampleUnitTest {
+    @Test
+    fun addition_isCorrect() {
+        assertEquals(4, 2 + 2)
+    }
+}
